+++ src/queue.h
@@ -10,8 +10,8 @@ typedef struct queuenode {
 } queuenode_t;
 
 typedef struct {
-  long          count;
-  long          currentIndex;
+  ssize_t       count;
+  ssize_t       currentIndex;
   queuenode_t   *iteratorNode;
   queuenode_t   *currentNode;
   queuenode_t   *head;
@@ -23,10 +23,10 @@ queue_t *queueAlloc (queueFree_t freeHook);
 void    queueFree (queue_t *q);
 void    queuePush (queue_t *q, void *data);
 void    *queueGetCurrent (queue_t *q);
-void    *queueGetByIdx (queue_t *q, long idx);
+void    *queueGetByIdx (queue_t *q, ssize_t idx);
 void    *queuePop  (queue_t *q);
-void    *queueRemoveByIdx (queue_t *q, long idx);
-long    queueGetCount (queue_t *q);
+void    *queueRemoveByIdx (queue_t *q, ssize_t idx);
+ssize_t queueGetCount (queue_t *q);
 void    queueStartIterator (queue_t *q);
 void    *queueIterateData (queue_t *q);
 void    *queueIterateRemoveNode (queue_t *q);
